refactor: utils 패키지 추가, ParseInputStringToNumbers 위치 변경
