+++ src/main/java/utils/ParseInputStringToNumbers.java
@@ -1,4 +1,4 @@
-package domain;
+package utils;
 
 import java.util.Arrays;
 import java.util.List;
